figured out the bad kerning in the statusbar
seriously what the hell?

This was debugged by running `fc-match -v roboto` on Arch and NixOS and
diffing the results. After trying overriding everything that was
different to be the Arch version, I determined that the bad kerning was
caused by the rgba property (????!), which I don't understand how it
could plausibly affect kerning.