new trace reformatter for exec traces to perfetto
This is deeply unsatisfying since the output is badly ordered and does not
maintain relationships properly, but due to the implicit parenting of events in
the legacy chrome trace file format, we can't do that much about that.

The solution to getting better output here is to switch to the perfetto trace
file format, for which I would probably have to write a haskell library to wrap
the protobufs. That would not be hard, but it would require "doing it".