feat: add optional sync of fpaths from $PATH
Currently, our $XDG_DATA_DIRs and $FPATH solution does not help with the
default behaviour of nix-shell [-p], which spawns a full shell,
obliviating some need for $FPATH modification post shell start, but
which does not export the `share/` of packages, (pending resolution of
https://github.com/NixOS/nix/issues/4254) and therefore will not be able
to provide completions at all.

This commit adds an optional workaround feature which attempts to detect
nix `share/` paths based on the bin directories on `$PATH`. It is
disabled by default, since this is abusing $PATH for something it is not
meant for. However, in practice, it is incredibly unlikely that this
will detect any directories that shouldn't be there anyway.