feat!: allow forcing use of builtin compinit instead of current compinit
Some plugins patch or even no-op the loaded compinit function and
therefore make it impossible to load new completions.

This commit adds a new `zstyle ':completion-sync:compinit:builtin-compinit' enabled`
which can be used to enable a workaround which temporarily replaces
compinit with the autoloaded one from the $FPATH, calls it and then
restores the original function.

This commit is breaking since the original plugin always used the
workaround, however this will likely do more harm than good and
therefore the default behaviour is to not use the workaround