fix(xdg): paths already on fpath during init get detected correctly
When initializing the `XDG_DATA_DIRS` sync we add all detected fpath
candidates from the totality of XDG_DATA_DIRS. Since this will detect a
lot of basic install paths (like the functions of the zsh install and
general system completions) it is important that we do not prepend all
paths, since this would massively boot the priority of paths that should
rather be fallback/builtin and easy to shadow on the fpath if needed.

The initialization for the XDG feature tries to detect this case by
checking whether the detected path is already present on the fpath, and
only adding it if it isn't. However the check was created faulty,
inserting extraneous double quotes which made all paths look
non-present. This commit fixes that